LayoutManager now is parent class to all layout views onResume for each layoutView is called from LayoutManager now
